Refactor code, reorganize files
