--- AUTHORS
@@ -1,19 +1,0 @@
-# This is the official list of 'w32' authors for copyright purposes.
-
-# Names should be added to this file as
-#   Name or Organization <email address>
-# The email address is not required for organizations.
-
-# Please keep the list sorted.
-
-# Contributors
-# ============
-
-Allen Dang <allengnr@gmail.com>
-Benny Siegert <bsiegert@gmail.com>
-Bruno Bigras <bigras.bruno@gmail.com>
-Daniel Joos
-Gerald Rosenberg <gerald.rosenberg@gmail.com>
-Liam Bowen <liambowen@gmail.com>
-Michael Henke
-Paul Maddox <paul.maddox@gmail.com>
