+++ com.go
@@ -1,7 +1,3 @@
-// Copyright 2010-2012 The W32 Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSE file.
-
 package w32
 
 import (
@@ -10,43 +6,86 @@ import (
 	"unsafe"
 )
 
-func MakeIntResource(id uint16) *uint16 {
-	return (*uint16)(unsafe.Pointer(uintptr(id)))
+type pIUnknownVtbl struct {
+	queryInterface uintptr
+	addRef         uintptr
+	release        uintptr
 }
 
-func LOWORD(dw uint32) uint16 {
-	return uint16(dw)
+type IUnknown struct {
+	lpVtbl *pIUnknownVtbl
 }
 
-func HIWORD(dw uint32) uint16 {
-	return uint16(dw >> 16 & 0xffff)
+func (u *IUnknown) QueryInterface(id *GUID) *IDispatch {
+	return ComQueryInterface(u, id)
 }
 
-func BoolToBOOL(value bool) BOOL {
-	if value {
-		return 1
-	}
-
-	return 0
+func (u *IUnknown) AddRef() int32 {
+	return ComAddRef(u)
 }
 
-func UTF16PtrToString(cstr *uint16) string {
-	if cstr != nil {
-		us := make([]uint16, 0, 256)
-		for p := uintptr(unsafe.Pointer(cstr)); ; p += 2 {
-			u := *(*uint16)(unsafe.Pointer(p))
-			if u == 0 {
-				return string(utf16.Decode(us))
-			}
-			us = append(us, u)
-		}
-	}
+func (u *IUnknown) Release() int32 {
+	return ComRelease(u)
+}
 
-	return ""
+type pIDispatchVtbl struct {
+	queryInterface   uintptr
+	addRef           uintptr
+	release          uintptr
+	getTypeInfoCount uintptr
+	getTypeInfo      uintptr
+	getIDsOfNames    uintptr
+	invoke           uintptr
+}
+
+type IDispatch struct {
+	lpVtbl *pIDispatchVtbl
+}
+
+func (d *IDispatch) QueryInterface(id *GUID) *IDispatch {
+	return ComQueryInterface((*IUnknown)(unsafe.Pointer(d)), id)
+}
+
+func (d *IDispatch) AddRef() int32 {
+	return ComAddRef((*IUnknown)(unsafe.Pointer(d)))
+}
+
+func (d *IDispatch) Release() int32 {
+	return ComRelease((*IUnknown)(unsafe.Pointer(d)))
+}
+
+func (d *IDispatch) GetIDsOfName(names []string) []int32 {
+	return ComGetIDsOfName(d, names)
+}
+
+func (d *IDispatch) Invoke(dispid int32, dispatch int16, params ...interface{}) *VARIANT {
+	return ComInvoke(d, dispid, dispatch, params...)
+}
+
+type pIStreamVtbl struct {
+	qeryInterface uintptr
+	addRef        uintptr
+	release       uintptr
+}
+
+type IStream struct {
+	lpVtbl *pIStreamVtbl
+}
+
+func (s *IStream) QueryInterface(id *GUID) *IDispatch {
+	return ComQueryInterface((*IUnknown)(unsafe.Pointer(s)), id)
+}
+
+func (s *IStream) AddRef() int32 {
+	return ComAddRef((*IUnknown)(unsafe.Pointer(s)))
+}
+
+func (s *IStream) Release() int32 {
+	return ComRelease((*IUnknown)(unsafe.Pointer(s)))
 }
 
 func ComAddRef(unknown *IUnknown) int32 {
-	ret, _, _ := syscall.Syscall(unknown.lpVtbl.pAddRef, 1,
+	ret, _, _ := syscall.Syscall(unknown.lpVtbl.addRef, 1,
 		uintptr(unsafe.Pointer(unknown)),
 		0,
 		0)
@@ -54,7 +93,7 @@ func ComAddRef(unknown *IUnknown) int32 {
 }
 
 func ComRelease(unknown *IUnknown) int32 {
-	ret, _, _ := syscall.Syscall(unknown.lpVtbl.pRelease, 1,
+	ret, _, _ := syscall.Syscall(unknown.lpVtbl.release, 1,
 		uintptr(unsafe.Pointer(unknown)),
 		0,
 		0)
@@ -63,7 +102,7 @@ func ComRelease(unknown *IUnknown) int32 {
 
 func ComQueryInterface(unknown *IUnknown, id *GUID) *IDispatch {
 	var disp *IDispatch
-	hr, _, _ := syscall.Syscall(unknown.lpVtbl.pQueryInterface, 3,
+	hr, _, _ := syscall.Syscall(unknown.lpVtbl.queryInterface, 3,
 		uintptr(unsafe.Pointer(unknown)),
 		uintptr(unsafe.Pointer(id)),
 		uintptr(unsafe.Pointer(&disp)))
@@ -79,7 +118,7 @@ func ComGetIDsOfName(disp *IDispatch, names []string) []int32 {
 	for i := 0; i < len(names); i++ {
 		wnames[i] = syscall.StringToUTF16Ptr(names[i])
 	}
-	hr, _, _ := syscall.Syscall6(disp.lpVtbl.pGetIDsOfNames, 6,
+	hr, _, _ := syscall.Syscall6(disp.lpVtbl.getIDsOfNames, 6,
 		uintptr(unsafe.Pointer(disp)),
 		uintptr(unsafe.Pointer(IID_NULL)),
 		uintptr(unsafe.Pointer(&wnames[0])),
@@ -104,7 +143,6 @@ func ComInvoke(disp *IDispatch, dispid int32, dispatch int16, params ...interfac
 	if len(params) > 0 {
 		vargs = make([]VARIANT, len(params))
 		for i, v := range params {
-			//n := len(params)-i-1
 			n := len(params) - i - 1
 			VariantInit(&vargs[n])
 			switch v.(type) {
@@ -175,7 +213,7 @@ func ComInvoke(disp *IDispatch, dispid int32, dispatch int16, params ...interfac
 	var ret VARIANT
 	var excepInfo EXCEPINFO
 	VariantInit(&ret)
-	hr, _, _ := syscall.Syscall9(disp.lpVtbl.pInvoke, 8,
+	hr, _, _ := syscall.Syscall9(disp.lpVtbl.invoke, 8,
 		uintptr(unsafe.Pointer(disp)),
 		uintptr(dispid),
 		uintptr(unsafe.Pointer(IID_NULL)),
@@ -198,4 +236,19 @@ func ComInvoke(disp *IDispatch, dispid int32, dispatch int16, params ...interfac
 	}
 	result = &ret
 	return
+}
+
+func UTF16PtrToString(cstr *uint16) string {
+	if cstr != nil {
+		us := make([]uint16, 0, 256)
+		for p := uintptr(unsafe.Pointer(cstr)); ; p += 2 {
+			u := *(*uint16)(unsafe.Pointer(p))
+			if u == 0 {
+				return string(utf16.Decode(us))
+			}
+			us = append(us, u)
+		}
+	}
+
+	return ""
 }
